scheduler/pyproject.toml
[project]
name = "scheduler"
version = "1.0.0"
description = "Scheduler"
requires-python = ">=3.12"
authors = [
    { name = "Bogdan Travin", email = "i.fatjuni@gmail.com" }
]
license = { text = "MIT" }
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "apscheduler[redis]==4.0.0a6",
    "aiohttp~=3.12.14",
    "pytz>=2025.2",
    "pydantic>=2.11.7",
    "python-jose[cryptography]~=3.3",
    "pydantic-settings>=2.10.1",
    "structlog>=25.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.hatch.build]
src-layout = false

[tool.hatch.build.targets.sdist]
include = [
    "/app",
    "/pyproject.toml",
]

[dependency-groups]
dev = [
    "ruff>=0.12.3",
]

[tool.ruff]
extend = "../pyproject.toml"
